+++ CHANGES.rst
@@ -1,6 +1,14 @@
 CHANGES
 =======
 
+0.3 (2017-xx-xx)
+----------------
+
+- Use distutils exceptions for errors.
+
+- Add rust version check for extension.
+
+
 0.2 (2017-03-08)
 ----------------